% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{Plum_runs}
\alias{Plum_runs}
\title{List the folders present in the current core directory.}
\usage{
Plum_runs(coredir = get("info")$coredir)
}
\arguments{
\item{coredir}{The directory where the Bacon runs reside. Defaults to \code{coredir="Plum_runs"}.}
}
\value{
A list of folders
}
\description{
Lists all folders located within the core's directory.
}
\details{
The directory is either "Plum_runs", "Cores" or a custom-named one.
}
\seealso{
\url{http://www.qub.ac.uk/chrono/blaauw/manualBacon_2.3.pdf}
}
\author{
Maarten Blaauw, J. Andres Christen
}
