% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plum.R
\docType{package}
\name{rplum}
\alias{rplum}
\title{rplum}
\description{
Plum is an approach to age-depth modelling that uses Bayesian statistics to reconstruct 
accumulation histories for 210Pb-dated deposits using prior information, and can combine 
210Pb, radiocarbon, and other dates in the chronologies. 
See Aquino et al. (2018) <doi:10.1007/s13253-018-0328-7>.
}
\author{
Maarten Blaauw <maarten.blaauw@qub.ac.uk> J. Andres Christen <jac@cimat.mx> Marco Aquino-Lopez <Marco.AquinoLopez@mu.ie>
}
